+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/oxfmt/settings/OxfmtSettings.kt
@@ -1,7 +1,12 @@
 package com.github.oxc.project.oxcintellijplugin.oxfmt.settings
 
 import com.github.oxc.project.oxcintellijplugin.ConfigurationMode
+import com.github.oxc.project.oxcintellijplugin.oxfmt.OxfmtPackage
+import com.github.oxc.project.oxcintellijplugin.oxfmt.OxfmtPackage.Companion.EMPTY_NODE_PACKAGE
 import com.github.oxc.project.oxcintellijplugin.oxfmt.settings.OxfmtSettingsState.Companion.DEFAULT_EXTENSION_LIST
+import com.intellij.javascript.nodejs.util.NodePackage
+import com.intellij.javascript.nodejs.util.NodePackageRef
+import com.intellij.lang.javascript.linter.JSNpmLinterState
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.SettingsCategory
 import com.intellij.openapi.components.SimplePersistentStateComponent
@@ -10,12 +15,13 @@ import com.intellij.openapi.components.Storage
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiElement
 import java.io.File
 
 @Service(Service.Level.PROJECT)
 @State(name = "OxfmtSettings", storages = [Storage("OxfmtSettings.xml")],
     category = SettingsCategory.TOOLS)
-class OxfmtSettings(private val project: Project) :
+class OxfmtSettings(private val project: Project) : JSNpmLinterState<OxfmtSettings>,
     SimplePersistentStateComponent<OxfmtSettingsState>(OxfmtSettingsState()) {
 
     var configurationMode: ConfigurationMode
@@ -55,6 +61,12 @@ class OxfmtSettings(private val project: Project) :
             state.fixAllOnSave = value
         }
 
+    var preferOxfmtCodeStyleSettings: Boolean
+        get() = state.preferOxfmtCodeStyleSettings
+        set(value) {
+            state.preferOxfmtCodeStyleSettings = value
+        }
+
     fun isEnabled(): Boolean {
         return configurationMode !== ConfigurationMode.DISABLED
     }
@@ -66,6 +78,20 @@ class OxfmtSettings(private val project: Project) :
         } else {
             false
         }
+    }
+
+    override fun getNodePackageRef(): NodePackageRef {
+        return NodePackageRef.create(getPackage(null))
+    }
+
+    override fun withLinterPackage(
+        nodePackageRef: NodePackageRef): OxfmtSettings {
+        return this
+    }
+
+    fun getPackage(context: PsiElement?): NodePackage {
+        return OxfmtPackage(project).getPackage(context?.containingFile?.virtualFile)
+               ?: EMPTY_NODE_PACKAGE
     }
 
     companion object {